Move app initialisation into a separate file
Now we can use the init script in the test bundle.
