+++ frontend/karma.conf.js
@@ -64,8 +64,6 @@ module.exports = function (config) {
       'bower_components/jquery-mockjax/jquery.mockjax.js',
       'tests/unit/tests/asset_functions.js',
 
-
-      '../app/assets/javascripts/bundles/openproject-core-app.js',
       '../app/assets/javascripts/bundles/openproject-tests.js'
     ],
